refactor: share base class for TOML and YAML
